添加上传剪裁附件

+++ src/main/java/com/xiaoshabao/blog/util/ImageUtils.java
@@ -91,7 +91,7 @@ public class ImageUtils {
 
         log.debug("origin with/height " + w + "/" + h);
 
-        int size = (int) Math.max(w, h);
+        int size = Math.max(w, h);
         int tow = w;
         int toh = h;
